+++ R/functions-msp.R
@@ -1,43 +1,54 @@
-##' @param f `character(1)` with the path to an msp file.
-##' 
-##' @param msLevel `numeric(1)` with the MS level. Default is 2.
-##' 
-##' @param ... Additional parameters, currently ignored.
-##'
-##' @importFrom S4Vectors DataFrame
-##'
-##' @importFrom IRanges NumericList
-##' 
-##' @author Laurent Gatto
-##' 
-##' @noRd
-
-if (FALSE) {
-    library(MSnbase)
-    library(microbenchmark)
-    f <- "/vol/bioinvindex/Submissions/MIC044/MTBLS1582/20203111526_spectra_SMpos.msp"
-    f <- "/vol/R/BioC/devel/Spectra/MassBank_MSMS_Pos_Rev173_vs1.msp"
-    f <- "/vol/R/BioC/devel/2018-03-09_14_10_01_pos_387850_nist_msms_HR.MSP"
-    f <- "/vol/R/BioC/devel/MsBackendMsp/inst/extdata/Spectrum.msp"
-    f <- "/vol/R/BioC/devel/MsBackendMsp/inst/extdata/Spectrum2.msp"
-    f <- "/vol/R/BioC/devel/MsBackendMsp/inst/extdata/msdial_pos.msp"
-    f <- "/vol/R/BioC/devel/MsBackendMsp/inst/extdata/first-export-LipidBlast.msp"
-    massbank <- .read_msp(f)
-    orgmassbank <- ReadMspFile(f)
-    
-    microbenchmark(.read_msp(f), times = 5)
-}
-
-
-.read_msp <- function(f, msLevel = 2L, ...) {
+#' @title Reading MSP files
+#'
+#' The `readMsp` function imports the data from a file in MGF format reading
+#' all specified fields and returning the data as a [DataFrame()].
+#'
+#' @param f `character(1)` with the path to an MSP file.
+#'
+#' @param msLevel `numeric(1)` with the MS level. Default is 2. This value will
+#'     be reported as the spectra's MS level **unless** the source MSP file
+#'     defines the MS level.
+#'
+#' @param mapping named `character` vector to rename MSP fields to spectra
+#'     variables (see [spectraVariableMapping()]). This allows to correctly
+#'     import also custom fields or data from files with different MSP
+#'     *flavors*.
+#'
+#' @param ... Additional parameters, currently ignored.
+#'
+#' @return
+#'
+#' A `DataFrame` with each row containing the data from one spectrum
+#' in the MSP file. m/z and intensity values are available in columns `"mz"`
+#' and `"intensity"` in a list representation.
+#'
+#' @export
+#'
+#' @importFrom S4Vectors DataFrame
+#'
+#' @importFrom IRanges NumericList
+#'
+#' @importFrom MsCoreUtils rbindFill
+#'
+#' @importFrom methods as
+#'
+#' @author Laurent Gatto, Steffen Neumann, Johannes Rainer
+#'
+#' @examples
+#'
+#' fls <- dir(system.file("extdata", package = "MsBackendMsp"),
+#'     full.names = TRUE, pattern = "msp$")[1L]
+#'
+#' readMsp(fls)
+readMsp <- function(f, msLevel = 2L,
+                    mapping = spectraVariableMapping(MsBackendMsp()), ...) {
     if (length(f) != 1L)
         stop("Please provide a single msp file.")
     
     msp <- scan(file = f, what = "",
                 sep = "\n", quote = "",
                 allowEscapes = FALSE,
-                quiet = TRUE
-                )
+                quiet = TRUE)
     
     ## Ignore comments
     cmts <- grep("^[#]", msp)
@@ -46,87 +57,118 @@ if (FALSE) {
 
     ## Find individual records
     begin <- grep("^NAME:", msp, ignore.case = TRUE)
-    stopifnot(!isEmpty(begin)) ## NO! DONT STOP ON ERROR! FIND OUT HOW EXCEPTIONS WORK!
-    
-    end <- c(begin[-1], length(msp))
+    end <- c(begin[-1] -1L, length(msp))
     
     n <- length(begin)
     sp <- vector("list", length = n)
 
-    for (i in seq(along = sp)) 
-        sp[[i]] <- .extract_msp_spectrum(msp[begin[i]:end[i]])
+    for (i in seq_along(sp))
+        sp[[i]] <- .extract_msp_spectrum(msp[begin[i]:end[i]],
+                                         mapping = mapping)
+    res <- DataFrame(rbindFill(sp))
 
-    res <- DataFrame(do.call(rbind, sp))
-
+    spv <- Spectra:::.SPECTRA_DATA_COLUMNS
+    spv <- spv[!names(spv) %in% c("mz", "intensity")]
     for (i in seq_along(res)) {
         if (all(lengths(res[[i]]) == 1))
             res[[i]] <- unlist(res[[i]])
+        if (any(col <- names(spv) == colnames(res)[i]))
+            res[[i]] <- as(res[[i]], spv[col][1])
     }
 
-    res$mz <- IRanges::NumericList(res$mz)
-    res$intensity <- IRanges::NumericList(res$intensity)
+    res$mz <- IRanges::NumericList(res$mz, compress = FALSE)
+    res$intensity <- IRanges::NumericList(res$intensity, compress = FALSE)
     res$dataOrigin <- f
-    res$msLevel <- as.integer(msLevel)
+    if (!any(colnames(res) == msLevel))
+        res$msLevel <- as.integer(msLevel)
     res
 }
 
-##' @param msp `character()` of lines defining a spectrum in msp
-##'     format.
-##' 
-##' @author Laurent Gatto
-##' 
-##' @importFrom stats setNames
-##'
-##' @noRd
-.extract_msp_spectrum <- function(msp) {
+#' @param msp `character()` of lines defining a spectrum in msp
+#'     format.
+#'
+#' @param mapping spectra variable mapping that allows renaming data fields.
+#' 
+#' @author Laurent Gatto, Johannes Rainer
+#' 
+#' @importFrom stats setNames
+#'
+#' @noRd
+.extract_msp_spectrum <- function(msp, mapping) {
     ## grep description
     desc.idx <- grep(":", msp)
     desc <- msp[desc.idx]
     spec <- msp[-desc.idx]
 
-    ms <- do.call(rbind, strsplit(spec, "[[:space:]]+"))
+    ms <- do.call(rbind, strsplit(sub("^(\\t|[[:space:]]+)", "", spec),
+                                  "[[:space:]]+"))
     mode(ms) <- "double"
 
     if (!length(ms))
         ms <- matrix(numeric(), ncol = 2L)
-    
-    if(length(ms) > 2)
-        ms <- ms[order(ms[, 1L]),]
+    else if (is.unsorted(ms[, 1L]))
+        ms <- ms[order(ms[, 1L]), ]
     
     r <- regexpr(":", desc, fixed = TRUE)
-    desc <- setNames(substring(desc, r + 2L, nchar(desc)), tolower(substring(desc, 1L, r - 1L)))
-    name <- unname(desc["name"])
-    accession <- unname(desc["db#"])
-    formula <- unname(desc["formula"])
-    inchikey <- unname(desc["inchikey"])
-    adduct <- unname(desc["precursor_type"])
-    exactmass <- as.numeric(unname(desc["exactmass"]))
-    comments <- unname(desc["comments"])
-    ion_mode <- unname(desc["ion_mode"])
+    desc <- setNames(substring(desc, r + 2L, nchar(desc)),
+                     substring(desc, 1L, r - 1L))
+    ## map fields to spectra variables
+    idx <- match(names(desc), mapping)
+    not_na <- !is.na(idx)
+    if (any(not_na))
+        names(desc)[not_na] <- names(mapping)[idx][not_na]
+
+    ## Handle eventually duplicated names -> list
+    if (anyDuplicated(names(desc))) {
+        res <- split(unname(desc), names(desc))
+        dups <- lengths(res) > 1L
+        dup_res <- res[dups]
+        res <- as.data.frame(res[!dups])
+        for (name in names(dup_res))
+            res <- do.call("$<-", list(res, name, unname(dup_res[name])))
+    } else res <- as.data.frame(as.list(desc))
     
-    ## select only values of interest and convert to numeric
-    voi <- c("retentiontime", "ionmode", "precursormz")
-    desc <- setNames(as.numeric(desc[voi]), voi)
-    desc[is.na(desc[voi])] <- 0L
-    
-    list(rtime = unname(desc["retentiontime"]),
-         precursorMz = unname(desc["precursormz"]),
-         precursorCharge = unname(as.integer(desc["ionmode"])),
-         mz = ms[, 1L],
-         intensity = ms[, 2L],
-         accession = accession,
-         name = name,
-         formula = formula,
-         inchikey = inchikey,
-         adduct = adduct,
-         exactmass = exactmass,
-         comments = comments,
-         ion_mode = ion_mode)
+    ## Ensure correct data type
+    ## polarity
+    if (any(have <- colnames(res) == "polarity"))
+        res[, have] <- .process_polarity(res[, have])
+    if (any(have <- colnames(res) == "msLevel"))
+        res[, have] <- .process_mslevel(res[, have])
+
+    res$mz <- list(ms[, 1L])
+    res$intensity <- list(ms[, 2L])
+    res
+}
+
+.process_polarity <- function(x, input = TRUE) {
+    if (input) {
+        if (grepl("^(p|\\+)", x, ignore.case = TRUE))
+            return(1L)
+        if (grepl("^(n|-)", x, ignore.case = TRUE))
+            return(0L)
+        -1L
+    } else {
+        xnew <- rep(NA_character_, length(x))
+        xnew[x == 1L] <- "Positive"
+        xnew[x == 0L] <- "Negative"
+        xnew
+    }
+}
+
+#' @param x value to be formatted
+#'
+#' @param input `logical(1)` whether the data is imported or exported.
+#'
+#' @noRd
+.process_mslevel <- function(x, input = TRUE) {
+    if (input)
+        as.integer(sub("ms", "", x, ignore.case = TRUE))
+    else paste0("MS", x)
 }
 
 #' @description
 #'
-#' Function to export a `Spectra` object in .msp format to `con`.
+#' Function to export a `Spectra` object in MSP format to `con`.
 #'
 #' @param x `Spectra`
 #'
@@ -134,87 +176,81 @@ if (FALSE) {
 #'
 #' @param mapping named `character` vector that maps from `spectraVariables`
 #'    (i.e. `names(mapping)`) to the variable name that should be used in the
-#'    MGF file.
+#'    MSP file.
 #'
-#' @author Michael Witting
+#' @param allVariables `logical(1)` whether all spectra variables in `x` should
+#'    be exported or only those that are listed in `mapping`. Note that if
+#'    `exportName = TRUE` a field *NAME* will be exported regardless of
+#'    `mapping`.
 #'
-#' @importMethodsFrom Spectra spectraVariables spectraNames spectraData
+#' @param exportName `logical(1)` whether a NAME field will always be exported
+#'    even if no such spectra variable is available in `x`.
+#' 
+#' @author Michael Witting, Johannes Rainer
 #'
-#' @noRd
-.export_msp <- function(x, con = stdout(), mapping = spectraVariableMapping()) {
-    
-    if (class(con) == "character" && file.exists(con)) {
-        
-        message("Overwriting ", con, "!")
-        unlink(con)
-        
-    }
-    
-    if (class(con)[1] == "character") {
-        con <- file(description = con, open = "at")
-        on.exit(close(con))
-    }
-    
-    # custom cat function for writing of content
-    .cat <- function(..., file = con, sep = " ", append = TRUE) {
-        cat(..., file = file, sep = sep, append = append)
-    }
-    
-    
-    # iterate over all spectra
-    for(i in 1:length(x)) {
-        
-        spv <- spectraVariables(x[i])
-        spd <- spectraData(x[i], spv[!(spv %in% c("dataOrigin", "dataStorage"))])
-        idx <- match(colnames(spd), names(mapping))
-        colnames(spd)[!is.na(idx)] <- mapping[idx[!is.na(idx)]]
-        
-        spp <- peaksData(x[i])
-        
-        # here list with stuff in right order
-        entries <- .getEntries()
-        
-        for(entry in entries) {
-            
-            #print(entry)
-            
-            if(entry %in% colnames(spd)) {
-                
-                value <- spd[entry][[1]]
-                
-                .cat(entry, value, "\n")
-            
-            }
-        }
-        
-        .cat("Num Peaks:", length(peaksData(x[i])[[1]][,1]), "\n")
-        
-        .cat(paste0(peaksData(x[i])[[1]][,1],
-                    " ",
-                    peaksData(x[i])[[1]][,2],
-                    collapse = "\n"))
-        
-        .cat("\n\n\n")
-    }
-}
-
-
-#' This list defines the order and fields used for export
+#' @importMethodsFrom Spectra spectraVariables spectraNames
+#'
+#' @importMethodsFrom Spectra peaksData spectraData
 #'
 #' @noRd
-.getEntries <- function() {
-    
-    c(
-        # record specific information
-        "NAME:",
-        "DB#:",
-        "INCHIKEY:",
-        "PRECURSORTYPE:",
-        "PRECURSORMZ:",
-        "RETENTIONTIME:",
-        "EXACTMASS:",
-        "FORMULA:"
-        )
+#'
+#' @examples
+#'
+#' spd <- DataFrame(msLevel = c(2L, 2L, 2L), rtime = c(1, 2, 3))
+#' spd$mz <- list(c(12, 14, 45, 56), c(14.1, 34, 56.1), c(12.1, 14.15, 34.1))
+#' spd$intensity <- list(c(10, 20, 30, 40), c(11, 21, 31), c(12, 22, 32))
+#'
+#' sps <- Spectra(spd)
+#'
+#' .export_msp(sps)
+#'
+#' ## Handling of variables with multiple entries
+#' sps$synonym <- list(c("a", "b"), "d", c("e", "f", "g"))
+#' .export_msp(sps)
+#'
+.export_msp <- function(x, con = stdout(),
+                        mapping = spectraVariableMapping(MsBackendMsp()),
+                        allVariables = TRUE, exportName = TRUE) {
+    spv <- spectraVariables(x)
+    spv <- spv[!(spv %in% c("dataOrigin", "dataStorage"))]
+    if (!allVariables)
+        spv <- spv[spv %in% names(mapping)]
+    spd <- spectraData(x, spv)
+    ## Process any known required data conversions
+    if (any(spv == "msLevel"))
+        spd$msLevel <- .process_mslevel(spd$msLevel, input = FALSE)
+    if (any(spv == "polarity"))
+        spd$polarity <- .process_polarity(spd$polarity, input = FALSE)
+    idx <- match(colnames(spd), names(mapping))
+    colnames(spd)[!is.na(idx)] <- mapping[idx[!is.na(idx)]]
+    ## Force variable NAME:
+    if (!any(tolower(colnames(spd)) == "name") && exportName)
+        spd$NAME <- seq_len(nrow(spd))
+    idx <- which(tolower(colnames(spd)) == "name")
+    if (length(idx)) {
+        idx <- idx[1L]
+        spd <- spd[, c(idx, seq_len(ncol(spd))[-idx])]
+    }
+
+    ## Determine which columns contain list-like data (i.e. multiple entries).
+    mult <- colnames(spd)[!vapply(spd, function(z) is.vector(z) & !is.list(z),
+                                  logical(1))]
+    for (m in mult) {
+        spd[, m] <- vapply(
+            spd[, m], function(z) paste0(z, collapse = paste0("\n", m, ": ")),
+            character(1))
+    }
+
+    tmp <- lapply(colnames(spd), function(z) {
+        paste0(z, ": ", spd[, z], "\n")
+    })
+
+    pks <- vapply(peaksData(x), function(z)
+        paste0("Num Peaks: ", nrow(z), "\n",
+               paste0(paste0(z[, 1], " ", z[, 2], "\n"), collapse = ""),
+               collapse = ""),
+        character(1))
+    tmp <- do.call(cbind, c(tmp, list(pks)))
+    tmp[grep(": NA\n", tmp, fixed = TRUE)] <- ""
+    writeLines(apply(tmp, 1, paste0, collapse = ""), con = con)
 }
-
-