+++ tests/testthat/test_functions-msp.R
@@ -1,31 +1,84 @@
-test_that(".read_msp works", {
-    fls <- dir(system.file("extdata", package = "MsBackendMsp"),
-               full.names = TRUE, pattern = "msp$")
+test_that("readMsp works", {
+    ## MS-DIAL
+    f <- system.file("extdata", "msdial_pos.msp", package = "MsBackendMsp")
 
-    expect_error(.read_msp(fls), "Please provide a single msp file.")
+    res <- readMsp(f, mapping = spectraVariableMapping(MsBackendMsp()))
+    expect_s4_class(res, "DataFrame")
+    expect_true(all(c("name", "precursorMz", "adduct", "rtime") %in%
+                    colnames(res)))
+    expect_s4_class(res$mz, "NumericList")
+    expect_s4_class(res$intensity, "NumericList")
+    expect_true(length(res$mz[1L]) == 1)
 
-    res1 <- .read_msp(fls[1])
-    res2 <- .read_msp(fls[2])
+    f <- system.file("extdata", "spectrum.msp", package = "MsBackendMsp")
+    f2 <- system.file("extdata", "spectrum2.msp", package = "MsBackendMsp")
 
-    # cns <- c("rtime", "scanIndex", "precursorMz", "precursorIntensity",
-    #          "precursorCharge", "mz", "intensity", "title", "dataOrigin",
-    #          "msLevel")
-    # expect_identical(sort(names(res1)), sort(cns))
-    # expect_identical(sort(names(res1)), sort(names(res2)))
-    # expect_true(is(res1$mz, "NumericList"))
-    # expect_true(is(res1$intensity, "NumericList"))
-    # expect_equal(length(res1$intensity[[1]]), length(res1$mz[[1]]))
-    # expect_equal(length(res1$intensity[[2]]), length(res1$mz[[2]]))
-    # expect_equal(length(res1$intensity[[3]]), length(res1$mz[[3]]))
-    # 
-    # expect_identical(res1$title,
-    #                  c("File193 Spectrum1719 scans: 2162",
-    #                    "File193 Spectrum1944 scans: 2406",
-    #                    "File193 Spectrum1968 scans: 2432"))
+    res <- readMsp(f, mapping = spectraVariableMapping(MsBackendMsp()))
+    res2 <- readMsp(f2, mapping = spectraVariableMapping(MsBackendMsp()))
+    res$dataOrigin <- "a"
+    res2$dataOrigin <- "a"
+    expect_equal(res, res2[1L, ])
+    
+    expect_error(readMsp(c(f, f2)), "Please provide a single msp file.")
 })
 
-test_that(".read_lipidblast_msp works", {
-    f <- system.file("extdata/small-export-LipidBlast.msp", package = "MsBackendMsp")
-    res1 <- .read_msp(f)
-    expect(all(dim(res1)==c(5,15)), "Dimensionality should be 5,15")
+test_that(".expect_msp_spectrum works", {
+    f <- system.file("extdata", "msdial_pos.msp", package = "MsBackendMsp")
+    x <- scan(file = f, what = "",
+              sep = "\n", quote = "",
+              allowEscapes = FALSE,
+              quiet = TRUE)
+    x <- x[15:22]
+
+    mapping <- spectraVariableMapping(MsBackendMsp())
+    res <- .extract_msp_spectrum(x, mapping = mapping)
+    expect_true(is.data.frame(res))
+    expect_true(nrow(res) == 1)
+
+    ## Duplicated values
+    x <- c(x, "synonym: a", "synonym: b", "other: 1", "other: 2", "other: 3")
+    res <- .extract_msp_spectrum(x, mapping = mapping)
+    expect_true(sum(colnames(res) == "other") == 1)
+    expect_true(sum(colnames(res) == "synonym") == 1)
+    expect_true(length(res$other[[1L]]) == 3)
+    expect_true(length(res$synonym[[1L]]) == 2)
+})
+
+test_that(".process_polarity works", {
+    expect_equal(.process_polarity("some"), -1L)
+    expect_equal(.process_polarity("Pos"), 1L)
+    expect_equal(.process_polarity("+"), 1L)
+    expect_equal(.process_polarity("Neg"), 0L)
+    expect_equal(.process_polarity("-"), 0L)
+
+    expect_equal(.process_polarity(1, input = FALSE), "Positive")
+    expect_equal(.process_polarity(0, input = FALSE), "Negative")
+    expect_equal(.process_polarity(2, input = FALSE), NA_character_)
+})
+
+test_that(".process_mslevel works", {
+    expect_equal(.process_mslevel("MS 2"), 2L)
+    expect_equal(.process_mslevel("ms1"), 1L)
+    expect_equal(.process_mslevel(1, input = FALSE), "MS1")
+})
+
+test_that(".export_msp works", {
+    spd <- DataFrame(msLevel = c(2L, 1L, 2L), rtime = c(1, 2, 3.2))
+    spd$mz <- list(c(12, 14, 45, 56), c(14.1, 34, 56.1), c(12.1, 14.15, 34.1))
+    spd$intensity <- list(c(10, 20, 30, 40), c(11, 21, 31), c(12, 22, 32))
+    sps <- Spectra(spd)
+
+    expect_output(.export_msp(sps[1], exportName = FALSE), "^msLevel: MS2")
+    expect_output(.export_msp(sps[1], exportName = FALSE), "Num Peaks: 4")
+
+    expect_output(.export_msp(sps[1], exportName = TRUE), "^NAME: 1")
+
+    sps$multi <- list("a", c("b", "c"), "e")
+    expect_output(.export_msp(sps), "multi: b\\nmulti: c")
+
+    map <- c(msLevel = "MSL", rtime = "TIME", multi = "OTHER")
+    expect_output(.export_msp(sps, mapping = map), "OTHER: b\\nOTHER: c")
+    expect_output(.export_msp(sps, mapping = map), "TIME: 1")
+    expect_output(.export_msp(sps, mapping = map), "TIME: 2")
+    expect_output(.export_msp(sps, mapping = map), "TIME: 3.2")    
 })