+++ NEWS.md
@@ -1,3 +1,27 @@
+# MsBackendMsp 0.98
+
+## Changes in 0.98.2
+
+- Vignette added.
+
+## Changes in 0.98.1
+
+- Improve export method to support multi-value fields and user-provided 
+  mappings.
+
+## Changes in 0.98.0
+
+- Add additional spectra variable mappings.
+- Call unit tests from the Spectra package.
+
+# MsBackendMsp 0.0
+
+## Changes in 0.0.5
+
+- Add possibility to define mapping between spectra variables and MSP data 
+  fields.
+- Support duplicated data fields in a single MSP entry.
+
 ## Changes in 0.0.4
 
 - Fix tests to match new columns in MSP parsing